refactor: extract cluster package from utils, move telemetry_parser
Group Kubernetes cluster operations (cluster_manager, node_selector,
pvc_utils, pattern_matcher) into a dedicated krkn_ai/cluster/ package.
Move telemetry_parser to chaos_engines where its only consumer lives.
utils/ now contains only generic cross-cutting utilities.
