+++ krkn_ai/cluster/__init__.py
@@ -1,0 +1,19 @@
+from krkn_ai.cluster.cluster_manager import ClusterManager
+from krkn_ai.cluster.node_selector import select_nodes, NodeSelectionResult
+from krkn_ai.cluster.pvc_utils import (
+    initialize_kubeconfig,
+    resolve_pod_name,
+    get_pvc_usage_percentage,
+)
+from krkn_ai.cluster.pattern_matcher import PatternMatcher, PatternValidationError
+
+__all__ = [
+    "ClusterManager",
+    "select_nodes",
+    "NodeSelectionResult",
+    "initialize_kubeconfig",
+    "resolve_pod_name",
+    "get_pvc_usage_percentage",
+    "PatternMatcher",
+    "PatternValidationError",
+]
