+++ krkn_ai/cluster/pattern_matcher.py
