+++ krkn_ai/models/scenario/scenario_io_hog.py
@@ -1,4 +1,4 @@
-from krkn_ai.utils.node_selector import select_nodes
+from krkn_ai.cluster import select_nodes
 from krkn_ai.models.scenario.base import Scenario
 from krkn_ai.models.scenario.parameters import (
     HogScenarioImageParameter,
